Implemented flow execution services.
Added unit tests for the logic to test run a flow in the web UI. In order to have the services tested all the running logic was extracted from the canvas component into a run service that calls the remote services and a RunnerService to orchestrate the calls to those services.

The remaining logic in the canvas component for running a flow is that of converting the internal info into the representation the services expect  and to update the diagram.

Also, methods in canvas component were reordered to group them by category and unused variables and methods were removed, hopefully this will ease future refactoring.