Fixes #351 - Substitute V1 API with V2 to get the Action data required to form the Diagram in Flow details page.
Refactored code. Added abstract class for Flows API calls. Changed name from Actions to Flows.

Implemented error handling to the service and rectified the api url to get the flow details from database.

Added error log to the console and added a spec for testing it.

Changed code such that the service can fit in with the actual application logic.