Quick support for contributions metadata properties rename.
This is an initial change to support #545. This change translates the renamed properties to the name the UI expects them to be as a quick solution to allow the engine code to use the new metadata without breaking compatibility with the UI.

This is only a temporal solution, the final solution should be to update the web UI server and client application to use the new property names.

This changes will execute the following on server side:

- When activities are loaded from the engine:
  - Renamed `input` property will be saved as `inputs` if present
  - Renamed `output` property will be saved as `outputs` if present
- When triggers are loaded from the engine:
  - Renamed `output` property will be saved as `outputs` if present