Fix unreachable trigger actionMappings.
This removes any unreachable trigger inputMappings after the flow schema is modified.

The issue happened when a flow defines input A, trigger maps to A and then input A is removed from the flow schema. When that happened the trigger mapping to A was still persisted in the handler but it is unreachable by the UI as flow schema doesn't define it anymore.

When the flogo.json was generated the mapping to A was still present.